+++ src/Ft3/PaymentHistory/PaymentOperation/TransferParam.cs
@@ -4,9 +4,9 @@ namespace Chromia.Postchain.Ft3
     {
         public readonly string AccountId;
         public readonly string AssetId;
-        public readonly float Amount;
+        public readonly int Amount;
 
-        public TransferParam(string accountId, string assetId, float amount)
+        public TransferParam(string accountId, string assetId, int amount)
         {
             this.AccountId = accountId;
             this.AssetId = assetId;
